Sunsetting v8008
s8008 and v8008 we incorrect implementations that ran into severe issues. s8008 made it all the way to hardware before I realized it was fundamentally incorrect.

v8008 had too much abstraction and shared logic. It was too complicated to debug and unwind.

b8008 is the newest attempt. We've move all components to legacy that relied on them. We've moved legacy projects and will revive them on a working 8008 - should we get there. This is a clean slate.